requirements.txt
requests
websockets
fastapi
uvicorn
pydantic
PyYAML
jinja2
